creat user queary test and tag comment test

+++ BlogMaster.Core/Command/Tag/DeleteTagCommand.cs
@@ -23,6 +23,10 @@ namespace BlogMaster.Core.Command.Tag;
           
 
             var result =await _tagRepository.GetTagById(request.Id);
+            if (result == null)
+            {
+                throw new Exception($"Tag not found{request.Id}");
+            }
             await _tagRepository.DeleteTag(result);
             return Unit.Value;
